+++ src/forms-opportunities/oppty-handlers/low-views-handler.js
@@ -22,7 +22,9 @@ import { DATA_SOURCES } from '../../common/constants.js';
  */
 // eslint-disable-next-line max-len
 export default async function createLowViewsOpportunities(auditUrl, auditDataObject, scrapedData, context, excludeForms = new Set()) {
-  const { dataAccess, log } = context;
+  const {
+    dataAccess, log, sqs, site, env,
+  } = context;
   const { Opportunity } = dataAccess;
 
   const auditData = JSON.parse(JSON.stringify(auditDataObject));
@@ -59,7 +61,7 @@ export default async function createLowViewsOpportunities(auditUrl, auditDataObj
         origin: 'AUTOMATION',
         title: 'Form has low views',
         description: 'The form has low views but the page containing the form has higher traffic',
-        tags: ['Form View'],
+        tags: ['Form Placement'],
         data: {
           ...opptyData,
           dataSources: [DATA_SOURCES.RUM, DATA_SOURCES.PAGE],
@@ -92,6 +94,24 @@ export default async function createLowViewsOpportunities(auditUrl, auditDataObj
         // eslint-disable-next-line no-await-in-loop
         await highPageViewsLowFormViewsOptty.save();
       }
+      log.info('sending message to mystique for high-page-views-low-form-views');
+      const mystiqueMessage = {
+        type: 'guidance:high-page-views-low-form-views',
+        siteId: auditData.siteId,
+        auditId: auditData.auditId,
+        deliveryType: site.getDeliveryType(),
+        time: new Date().toISOString(),
+        data: {
+          url: opportunityData.data.form,
+          form_source: opportunityData.data.formsource,
+          cta_text: '', // This will be available after merging the changes for scraping form CTA text
+          cta_source: '', // This will be available after merging the changes for scraping form CTA text
+        },
+      };
+
+      // eslint-disable-next-line no-await-in-loop
+      await sqs.sendMessage(env.QUEUE_SPACECAT_TO_MYSTIQUE, mystiqueMessage);
+      log.info(`forms opportunity high page views low form views sent to mystique: ${JSON.stringify(mystiqueMessage)}`);
     }
   } catch (e) {
     log.error(`Creating Forms opportunity for high page views low form views for siteId ${auditData.siteId} failed with error: ${e.message}`, e);