+++ test/audits/forms/oppty-handlers/low-views-oppoty-handler.test.js
@@ -65,12 +65,47 @@ describe('createLowFormViewsOpportunities handler method', () => {
       dataAccess: dataAccessStub,
       env: {
         S3_SCRAPER_BUCKET_NAME: 'test-bucket',
+        QUEUE_SPACECAT_TO_MYSTIQUE: 'test-queue',
       },
       site: {
         getId: sinon.stub().returns('test-site-id'),
+        getDeliveryType: sinon.stub().returns('eds'),
+      },
+      sqs: {
+        sendMessage: sinon.stub().resolves({}),
       },
     };
     auditData = testData.lowFormviewsAuditData;
+  });
+
+  it('should send message to mystique', async () => {
+    dataAccessStub.Opportunity.allBySiteIdAndStatus.resolves([highPageViewsLowFormViewsOptty]);
+    await createLowViewsOpportunities(auditUrl, auditData, undefined, context);
+    const expectedMessage = {
+      type: 'guidance:high-page-views-low-form-views',
+      siteId: 'site-id',
+      auditId: 'audit-id',
+      deliveryType: 'eds',
+      data: {
+        url: 'https://www.surest.com/existing-opportunity',
+        form_source: '',
+        cta_text: '',
+        cta_source: '',
+      },
+    };
+    expect(context.sqs.sendMessage).to.be.calledWith(
+      'test-queue',
+      sinon.match((actual) => (
+        actual.type === expectedMessage.type
+        && actual.siteId === expectedMessage.siteId
+        && actual.auditId === expectedMessage.auditId
+        && actual.data.url === expectedMessage.data.url
+        && actual.deliveryType === expectedMessage.deliveryType
+        && actual.data.form_source === expectedMessage.data.form_source
+        && actual.data.cta_text === expectedMessage.data.cta_text
+        && actual.data.cta_source === expectedMessage.data.cta_source
+      ), 'matches expected message excluding timestamp'),
+    );
   });
 
   it('should create new high page views low form views opportunity', async () => {
@@ -83,7 +118,7 @@ describe('createLowFormViewsOpportunities handler method', () => {
       title: 'Form has low views',
       description: 'The form has low views but the page containing the form has higher traffic',
       tags: [
-        'Form View',
+        'Form Placement',
       ],
       data: {
         form: 'https://www.surest.com/high-page-low-form-view',